メッセージコンポーネント共有化
